pyproject.toml
[tool.poetry]
name = "api-for-books-library"
version = "0.1.0"
description = ""
authors = ["KonstantinMikheev <mikhey17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-stubs = "^5.1.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
